requirements.txt
# Web server packages
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
python-multipart>=0.0.20

# LangChain packages
langchain>=0.0.267
langchain-core>=0.0.10
langchain-community>=0.0.10
langchain-huggingface>=0.0.6
langchain-chroma>=0.0.3
langchain-ollama>=0.0.3
langchain-openai>=0.0.6
langgraph>=0.0.20

# Vector Database and Search
chromadb>=0.4.18
elasticsearch>=8.10.0

# OpenAI API
openai>=1.3.0

# Other dependencies
sentence-transformers>=2.2.2
minio==7.2.0
uuid==1.30
pypdf2>=3.0.0
docx2txt>=0.8.0
langchain-text-splitters>=0.0.1

# Database dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
alembic>=1.10.0
uuid==1.30
